+++ src/Engine/Base/Camera.cpp
@@ -30,7 +30,7 @@ Camera::Camera(glm::vec3 position, glm::vec3 up, float yaw, float pitch)
 //    }
 
     // returns the view matrix calculated using Euler Angles and the LookAt Matrix
-    glm::mat4 Camera::GetViewMatrix()
+    glm::mat4 Camera::GetViewMatrix() const
     {
         return glm::lookAt(Position, Position + Front, Up);
     }